% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_helpr.R
\name{description_deltaprior}
\alias{description_deltaprior}
\title{Generate description for the prior specification on the number of dispersal routes, \eqn{\Delta}.}
\usage{
description_deltaprior(
  states_num,
  symmetry = T,
  delta_prior = c("Poisson", "Uniform", "Beta-Binomial"),
  poisson_default = T,
  poisson_lambda = 0,
  alpha_beta = 1,
  beta_beta = 1
)
}
\arguments{
\item{states_num}{Number of discrete states of the study trait.}

\item{symmetry}{Whether the specified geographic model is symmetric (true) or asymmetric (false)}

\item{delta_prior}{Which of the three prior options to put on \eqn{\Delta}, including
"Poisson": a(n offset) Poisson distribution (the default option);
"Uniform": a uniform distribution between zero and the maximum \eqn{\Delta} (when all the dispersal routes exist), and;
"Beta-Binomial": a Beta-Binomial distribution.}

\item{poisson_default}{Whether to specify the default Poisson prior in BEAUti.}

\item{poisson_lambda}{The rate parameter (\eqn{\lambda}) of the Poisson distribution.}

\item{alpha_beta}{Parameter \eqn{\alpha} of the Beta-Binomial distribution.}

\item{beta_beta}{Parameter \eqn{\beta} of the Beta-Binomial distribution.}
}
\value{
an R markdown paragraph describing the specified prior on the number of dispersal routes, \eqn{\Delta}.
}
\description{
Generate description for the prior specification on the number of dispersal routes, \eqn{\Delta}.
}
\examples{
description_deltaprior(5L, T, "Poisson", T)
}
