% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_helpr.R
\name{description_muprior}
\alias{description_muprior}
\title{Generate description for the prior specification on the average dispersal rate, \eqn{\mu}.}
\usage{
description_muprior(
  mu_prior = c("CTMC rate-ref (BEAST default)", "Hierarchical Exponential",
    "Empirical-Informed Exponential"),
  tree_length = 1,
  tree_num = 1L,
  hierachexp_mu95interval = NULL,
  hierachexp_alphaofgamma = 0.5,
  hierachexp_dispersaleventsnummean = 1,
  hierachexp_dispersaleventsnum95interval = NULL,
  empinformed_dispersaleventsnummean = 1,
  parsimonyscore_quantile = 0.5
)
}
\arguments{
\item{mu_prior}{Which of the three prior options to put on \eqn{\mu}, including
"CTMC rate-ref (BEAST default)":
"Hierarchical Exponential":
"Empirical-Informed Exponential":}

\item{tree_length}{Length of the phylogenetic tree (sum of all the branch lengths, in unit of time).}

\item{tree_num}{Number of phylogenetic trees to condition on (1 when the tree is fixed).}

\item{hierachexp_mu95interval}{\eqn{95\%} prior interval of the hierarchical Exponential distribution.}

\item{hierachexp_alphaofgamma}{Shape and rate parameter of the Gamma hyperprior of the Hierarchical Exponential prior.}

\item{hierachexp_dispersaleventsnummean}{Prior mean of the number of dispersal events over the phylogeny under the Hierarchical Exponential prior.}

\item{hierachexp_dispersaleventsnum95interval}{\eqn{95\%} prior interval of the number of dispersal events over the phylogeny under the hierarchical Exponential prior.}

\item{empinformed_dispersaleventsnummean}{Prior mean of the number of dispersal events over the phylogeny under the Empirical-Informed Exponential prior.}

\item{parsimonyscore_quantile}{The quantile of the prior distribution of the number of dispersal events to the put the parsimony score at.}
}
\value{
an R markdown paragraph describing the specified prior on the average dispersal rate, \eqn{\mu}.
}
\description{
Generate description for the prior specification on the average dispersal rate, \eqn{\mu}.
}
\examples{
description_muprior("CTMC rate-ref (BEAST default)", 1000, 1)
}
