% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorpredictive_functions.R
\name{draw_nevents_conditional}
\alias{draw_nevents_conditional}
\title{Draw the number of events conditioning on the start and end state given the underlying CTMC 
(implementing equation 2.9 from Hobolth and Stone, 2009)}
\usage{
draw_nevents_conditional(node_ages, Q, Q_ages = NULL, node_state_indices, Pab)
}
\arguments{
\item{node_ages}{A vector containing the start (first element) and end ages (second element) of this duration}

\item{Q}{An instantaneous-rate matrix (or a list of matrices for a piecewise constant discrete-geographic model) characterizes the CTMC}

\item{Q_ages}{A vector containing boundaries of time intervals for a piecewise constant geographic model (NULL means a constant model);
the length of this vector should be one shorter than the Q-matrix list, sorted from oldest to youngest.}

\item{node_state_indices}{State indices (i.e., the indices to use to fecth the states from the states vector) pf the start and end nodes}

\item{Pab}{The transition probability between the start and end states}
}
\value{
An integer that is the sampled number of events
}
\description{
Draw the number of events conditioning on the start and end state given the underlying CTMC 
(implementing equation 2.9 from Hobolth and Stone, 2009)
}
\keyword{internal}
