% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorpredictive_functions.R
\name{fill_DTMCmats}
\alias{fill_DTMCmats}
\title{Compute the uniformized matrix (R) of Q and its exponentiations (R^2, R^3, ..., R^n)
Here the vector of uniformized matrices in stored globally so we check if the provided n (number of events)
exceeds the maximum exponent minus 1 (as the first element is an identity matrix)
and only append the additional exponentiation(s) when it exceeds}
\usage{
fill_DTMCmats(Q, n, indices = NULL)
}
\arguments{
\item{Q}{An instantaneous-rate matrix (or a list of matrices for a piecewise constant geographic model) characterizes the CTMC}

\item{n}{Number of events (which is also the largest exponent to compute in this step)}

\item{indices}{Which interval we need to update (if not provided then update all)}
}
\description{
Compute the uniformized matrix (R) of Q and its exponentiations (R^2, R^3, ..., R^n)
Here the vector of uniformized matrices in stored globally so we check if the provided n (number of events)
exceeds the maximum exponent minus 1 (as the first element is an identity matrix)
and only append the additional exponentiation(s) when it exceeds
}
\keyword{internal}
