% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorpredictive_helpr.R
\name{historySimulator_teststatisticsComputer}
\alias{historySimulator_teststatisticsComputer}
\title{Simulate discrete-character histories using the inferred posterior distribution and compute posterior-predictive summary statistics using the simulated histories}
\usage{
historySimulator_teststatisticsComputer(
  states_dat,
  taxon_name,
  discrete_trait_name,
  tree_path,
  log_path,
  simsamples_num = 1000L,
  simulated_outpath
)
}
\arguments{
\item{states_dat}{A data-frame object containing at least two columns: one for the name of each tip (column name specified by \code{taxon_name}), 
and the other for the trait state of each tip (column name specified by \code{discrete_trait_name}).}

\item{taxon_name}{Name of the column containing tip names.}

\item{discrete_trait_name}{Name of the column containing tip states.}

\item{tree_path}{File path of the posterior distribution of trees that will be read in to simulated over.}

\item{log_path}{File path of the posterior distribution of parameters whose values will be used in the simulation.}

\item{simsamples_num}{Number of histories to simulate.}

\item{simulated_outpath}{File path to write the simulated data to.}
}
\value{
A data-frame object that contains the simulated and observed values of the two summary statistics
}
\description{
Simulate discrete-character histories using the inferred posterior distribution and compute posterior-predictive summary statistics using the simulated histories
}
\examples{

}
