% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorpredictive_functions.R
\name{sim_history_branch_conditional}
\alias{sim_history_branch_conditional}
\title{Simulate history of discrete-character changes over a branch (which may comprise one or more pieces delimited by time interval bounds) 
conditioning on the start and end states}
\usage{
sim_history_branch_conditional(
  node_ages,
  Q,
  Q_ages = NULL,
  node_state_indices,
  states
)
}
\arguments{
\item{node_ages}{Ages of the two nodes defining this branch}

\item{Q}{An instantaneous-rate matrix (or a list of matrices for a piecewise constant geographic model) characterizes the CTMC}

\item{Q_ages}{A vector containing boundaries of time intervals for a piecewise constant geographic model (NULL means a constant model);
the length of this vector should be one shorter than the Q-matrix list, sorted from oldest to youngest.}

\item{node_state_indices}{State indices (i.e., the indices to use to fecth the states from the states vector) of the start and end nodes}

\item{states}{States of the discrete character}
}
\value{
A map object (as in the simmap data structure) that contains the simulated history (as a series of events) over this branch
}
\description{
Simulate history of discrete-character changes over a branch (which may comprise one or more pieces delimited by time interval bounds) 
conditioning on the start and end states
}
