% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorpredictive_functions.R
\name{sim_history_piece_conditional}
\alias{sim_history_piece_conditional}
\title{Simulate history of discrete-character changes over a duration of time (i.e., a piece of a branch that the character evolved under a constant model)
conditioning on the start and end states (implementing algorithm #5 of Hobolth and Stone, 2009)}
\usage{
sim_history_piece_conditional(
  node_ages,
  Q,
  Q_ages = NULL,
  P = NULL,
  node_state_indices
)
}
\arguments{
\item{node_ages}{Start and end ages of this duration}

\item{Q}{An instantaneous-rate matrix (or a list of matrices for a piecewise constant geographic model) characterizes the CTMC}

\item{Q_ages}{A vector containing boundaries of time intervals for a piecewise constant geographic model (NULL means a constant model);
the length of this vector should be one shorter than the Q-matrix list, sorted from oldest to youngest.}

\item{P}{The transition probability matrix computed from the Q matrix (if not provided then it will be computed in this function)}

\item{node_state_indices}{State indices (i.e., the indices to use to fecth the states from the states vector) pf the start and end nodes}
}
\value{
A list containing two vectors, one for the end state of each event over this duration and other for the age of each event
}
\description{
Simulate history of discrete-character changes over a duration of time (i.e., a piece of a branch that the character evolved under a constant model)
conditioning on the start and end states (implementing algorithm #5 of Hobolth and Stone, 2009)
}
\keyword{internal}
