% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_helpr.R
\name{tex_analysis}
\alias{tex_analysis}
\title{Combine the descriptions on different types of analyses to generate the Analysis subsection of the Methods section.}
\usage{
tex_analysis(
  further_analysis = "",
  posterioranalysis_text = "",
  summarystats_text = "",
  powerposterior_text = "",
  prioranalysis_text = "",
  dcanalysis_text = "",
  format = c("HTML", "Latex", "Markdown")
)
}
\arguments{
\item{further_analysis}{The type of analysis selected to run, including:
\itemize{
  \item "": regular posterior analysis;
  \item "Marginal likelihood estimation": power-posterior analysis to estimate marginal likelihood (after regular posterior analysis);
  \item "Under prior": analysis to infer the joint prior distribution, and;
  \item "Data cloning": data-cloned posterior analysis.
}}

\item{posterioranalysis_text}{Description of the MCMC simulation inferring the posterior distribution.}

\item{summarystats_text}{Description of the geographic history inference.}

\item{powerposterior_text}{Description of the power-posterior analysis.}

\item{prioranalysis_text}{Description of MCMC simulations where we infer the joint prior distribution.}

\item{dcanalysis_text}{Description of the MCMC simulations inferring the data-cloned posterior distribution.}

\item{format}{One of "HTML", "Latex", or "Markdown".}
}
\value{
paragraphs (in the selected format) describing the analyses of the study
}
\description{
Combine the descriptions on different types of analyses to generate the Analysis subsection of the Methods section.
}
\examples{
posterioranalysis_text <- tex_posterioranalysis(1000000, 1000, 2)
summarystats_text <- tex_summarystats("Stochastic mapping (complete history, simulation-based)")
powerposterior_text <- tex_powerposterior(64, 500000, 1000, 0.3, 2)
tex_analysis(further_analysis = "Marginal likelihood estimation", posterioranalysis_text = posterioranalysis_text, summarystats_text = summarystats_text, powerposterior_text = powerposterior_text)
}
\keyword{internal}
