% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_helpr.R
\name{tex_bayesianinference}
\alias{tex_bayesianinference}
\title{Generate the description (in the selected format) of the Bayesian inference procedure used in the study.}
\usage{
tex_bayesianinference(
  empiricaltree_mh = T,
  tree_num = 1,
  bssvs = T,
  format = c("HTML", "Latex", "Markdown"),
  render_citation = T
)
}
\arguments{
\item{empiricaltree_mh}{Whether treating the provided distribution of trees as part of the Markov chain (default true) or averaging over the distribution uniformly.}

\item{tree_num}{Number of phylogenetic trees to condition on (1 when the tree is fixed).}

\item{bssvs}{Whether Bayesian Stochastic Search Variable Selection (BSSVS) is used (default true) in the discrete-geographic model.}

\item{format}{One of "HTML", "Latex", or "Markdown".}

\item{render_citation}{Whether the reference list will be rendered at the end of the methods document (default true; false when just showing the description in the HTML panel).}
}
\value{
a paragraph (in the selected format) describing the Bayesian inference procedure used in the study.
}
\description{
Generate the description (in the selected format) of the Bayesian inference procedure used in the study.
}
\examples{
tex_bayesianinference(TRUE, 1, TRUE)
}
\keyword{internal}
