% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_helpr.R
\name{tex_dcanalysis}
\alias{tex_dcanalysis}
\title{Generate the description (in the selected format) of the MCMC simulations inferring the data-cloned posterior distribution.}
\usage{
tex_dcanalysis(
  lheat = 1,
  mcmc_chainlength = 1,
  mcmc_samplingfreq = 1,
  mcmc_numreplicates = 2,
  format = c("HTML", "Latex", "Markdown"),
  render_citation = T
)
}
\arguments{
\item{lheat}{Number of data clones}

\item{mcmc_chainlength}{Number of MCMC generations to run the analysis}

\item{mcmc_samplingfreq}{Every number of MCMC generations to log to the output}

\item{mcmc_numreplicates}{Number of MCMC replicates to run}

\item{format}{One of "HTML", "Latex", or "Markdown".}
}
\value{
a paragraph (in the selected format) describing the MCMC simulations inferring the data-cloned posterior distribution.
}
\description{
Generate the description (in the selected format) of the MCMC simulations inferring the data-cloned posterior distribution.
}
\examples{
tex_dcanalysis(5, 1000000, 1000, 2)
}
\keyword{internal}
