% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_helpr.R
\name{tex_deltaprior}
\alias{tex_deltaprior}
\title{Generate the description (in the selected format) of the prior on the number of dispersal routes, \eqn{\Delta}, specified in the study.}
\usage{
tex_deltaprior(
  states_num,
  symmetry = T,
  delta_prior = c("Poisson", "Uniform", "Beta-Binomial"),
  poisson_default = T,
  poisson_lambda = 0,
  alpha_beta = 1,
  beta_beta = 1,
  format = c("HTML", "Latex", "Markdown"),
  render_citation = T
)
}
\arguments{
\item{states_num}{Number of discrete states of the study trait.}

\item{symmetry}{Whether the discrete-geographic mode is symmetric (default true) or asymmetric.}

\item{delta_prior}{Which of the three prior options to put on \eqn{\Delta}, including:
\itemize{
  \item "Poisson": a(n offset) Poisson distribution (the default option);
  \item "Uniform": a uniform distribution between zero and the maximum \eqn{\Delta} (when all the dispersal routes exist), and;
  \item "Beta-Binomial": a Beta-Binomial distribution.
}}

\item{poisson_default}{Whether to specify the default Poisson prior in BEAUti.}

\item{poisson_lambda}{The rate parameter (\eqn{\lambda}) of the Poisson distribution.}

\item{alpha_beta}{Parameter \eqn{\alpha} of the Beta-Binomial distribution.}

\item{beta_beta}{Parameter \eqn{\beta} of the Beta-Binomial distribution.}

\item{format}{One of "HTML", "Latex", or "Markdown".}

\item{render_citation}{Whether the reference list will be rendered at the end of the methods document (default true; false when just showing the description in the HTML panel).}
}
\value{
a paragraph (in the selected format) describing the prior on the number of dispersal routes specified in the study.
}
\description{
Generate the description (in the selected format) of the prior on the number of dispersal routes, \eqn{\Delta}, specified in the study.
}
\examples{
tex_deltaprior(states_num = 5L, symmetry = TRUE, delta_prior = "Poisson", poisson_default = TRUE, poisson_lambda = log(2))
}
\keyword{internal}
