% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_helpr.R
\name{tex_model}
\alias{tex_model}
\title{Generate the description (in the selected format) of the model specified in the study.}
\usage{
tex_model(
  bssvs = T,
  symmetry = T,
  format = c("HTML", "Latex", "Markdown"),
  render_citation = T
)
}
\arguments{
\item{bssvs}{Whether Bayesian Stochastic Search Variable Selection (BSSVS) is used (default true) in the discrete-geographic model.}

\item{symmetry}{Whether the discrete-geographic mode is symmetric (default true) or asymmetric.}

\item{format}{One of "HTML", "Latex", or "Markdown".}

\item{render_citation}{Whether the reference list will be rendered at the end of the methods document (default true; false when just showing the description in the HTML panel)}
}
\value{
a paragraph (in the selected format) describing the model specified in the study.
}
\description{
Generate the description (in the selected format) of the model specified in the study.
}
\examples{
tex_model(TRUE, TRUE, "Markdown", TRUE)
}
\keyword{internal}
