% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_helpr.R
\name{tex_muprior}
\alias{tex_muprior}
\title{Generate the description (in the selected format) of the prior on the average dispersal rate, \eqn{\mu}, specified in the study.}
\usage{
tex_muprior(
  mu_prior = c("CTMC rate-ref (BEAST default)", "Hierarchical Exponential",
    "Empirical-Informed Exponential"),
  tree_length = 1,
  tree_num = 1L,
  hierachexp_alphaofgamma = 0.5,
  hierachexp_dispersaleventsnummean = 1,
  empinformed_dispersaleventsnummean = 1,
  parsimony_score = 1,
  format = c("HTML", "Latex", "Markdown"),
  render_citation = T
)
}
\arguments{
\item{mu_prior}{Which of the three prior options to put on \eqn{\mu}, including:
\itemize{
  \item "CTMC rate-ref (BEAST default)": default prior recommended by BEAUti on the average dispersal rate;
  \item "Hierarchical Exponential": an exponential prior whose mean is under a hyperprior and will be inferred from the data, and;
  \item "Empirical-Informed Exponential": an exponential prior whose mean is informed by prior knowledge of the study system.
}}

\item{tree_length}{Length of the phylogenetic tree (sum of all the branch lengths, in unit of time).}

\item{tree_num}{Number of phylogenetic trees to condition on (1 when the tree is fixed).}

\item{hierachexp_alphaofgamma}{Shape and rate parameter of the Gamma hyperprior of the Hierarchical Exponential prior.}

\item{hierachexp_dispersaleventsnummean}{Prior mean of the number of dispersal events over the phylogeny under the Hierarchical Exponential prior.}

\item{empinformed_dispersaleventsnummean}{Prior mean of the number of dispersal events over the phylogeny under the Empirical-Informed Exponential prior.}

\item{parsimony_score}{Parsimony number for the discrete-geographic data given the provided tree.}

\item{format}{One of "HTML", "Latex", or "Markdown".}

\item{render_citation}{Whether the reference list will be rendered at the end of the methods document (default true; false when just showing the description in the HTML panel).}
}
\value{
a paragraph (in the selected format) describing the prior on the average dispersal rate specified in the study.
}
\description{
Generate the description (in the selected format) of the prior on the average dispersal rate, \eqn{\mu}, specified in the study.
}
\examples{
tex_muprior(mu_prior = "CTMC rate-ref (BEAST default)", tree_length = 1000, tree_num = 1, parsimony_score = 500L)
}
\keyword{internal}
