% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_helpr.R
\name{tex_posterioranalysis}
\alias{tex_posterioranalysis}
\title{Generate the description (in the selected format) of the MCMC simulation inferring the posterior distribution.}
\usage{
tex_posterioranalysis(
  mcmc_chainlength = 1,
  mcmc_samplingfreq = 1,
  mcmc_numreplicates = 2,
  format = c("HTML", "Latex", "Markdown")
)
}
\arguments{
\item{mcmc_chainlength}{Number of MCMC generations to run the analysis}

\item{mcmc_samplingfreq}{Every number of MCMC generations to log to the output}

\item{mcmc_numreplicates}{Number of MCMC replicates to run}

\item{format}{One of "HTML", "Latex", or "Markdown".}
}
\value{
a paragraph (in the selected format) describing the MCMC simulation inferring the posterior distribution.
}
\description{
Generate the description (in the selected format) of the MCMC simulation inferring the posterior distribution.
}
\examples{
tex_posterioranalysis(1000000, 1000, 2)
}
\keyword{internal}
