% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_helpr.R
\name{tex_powerposterior}
\alias{tex_powerposterior}
\title{Generate the description (in the selected format) of the power-posterior analysis used in the study.}
\usage{
tex_powerposterior(
  ml_numstones = 0,
  ml_chainlengthperstone = 0,
  ml_samplingfreq = 0,
  ml_alphaofbeta = 0.3,
  mcmc_numreplicates = 2,
  format = c("HTML", "Latex", "Markdown"),
  render_citation = T
)
}
\arguments{
\item{ml_numstones}{Number of powers ("stones") to use in the power-posterior analysis.}

\item{ml_chainlengthperstone}{Number of generations to run per stone.}

\item{ml_samplingfreq}{Every number of generations to log to the output.}

\item{ml_alphaofbeta}{Alpha of the Beta distribution that decides where to put the stones.}

\item{mcmc_numreplicates}{Number of MCMC replicates to run.}

\item{format}{One of "HTML", "Latex", or "Markdown".}

\item{render_citation}{Whether the reference list will be rendered at the end of the methods document (default true; false when just showing the description in the HTML panel).}
}
\value{
A paragraph (in the selected format) describing the power-posterior analysis used in the study.
}
\description{
Generate the description (in the selected format) of the power-posterior analysis used in the study.
}
\examples{
tex_powerposterior(ml_numstones = 64, ml_chainlengthperstone = 500000, ml_samplingfreq = 1000, ml_alphaofbeta = 0.3, mcmc_numreplicates = 2)
}
\keyword{internal}
