% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_helpr.R
\name{tex_prior}
\alias{tex_prior}
\title{Combine the descriptions on the two priors to generate the Prior subsection of the Methods section.}
\usage{
tex_prior(
  deltaprior_text,
  muprior_text,
  format = c("HTML", "Latex", "Markdown")
)
}
\arguments{
\item{deltaprior_text}{Description of the prior on the number of dispersal routes, \eqn{\Delta}.}

\item{muprior_text}{Description of the prior on the average dispersal rate, \eqn{\mu}.}

\item{format}{One of "HTML", "Latex", or "Markdown".}
}
\value{
paragraphs (in the selected format) describing prior specification in the study
}
\description{
Combine the descriptions on the two priors to generate the Prior subsection of the Methods section.
}
\examples{
deltaprior_text <- tex_deltaprior(states_num = 5L, symmetry = TRUE, delta_prior = "Poisson", poisson_default = TRUE, format = "Markdown")
muprior_text <- tex_muprior(mu_prior = "CTMC rate-ref (BEAST default)", tree_length = 1000, tree_num = 1L, parsimony_score = 500L)
tex_prior(deltaprior_text, muprior_text, "HTML")
}
