% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_helpr.R
\name{tex_summarystats}
\alias{tex_summarystats}
\title{Generate the description (in the selected format) of the geographic history inference in the study.}
\usage{
tex_summarystats(
  do_stochasticmapping = "",
  markovjumps_total = T,
  markovjumps_pairwise = T,
  format = c("HTML", "Latex", "Markdown"),
  render_citation = T
)
}
\arguments{
\item{do_stochasticmapping}{Whether to perform stochastic mapping to simulate full histories of the discrete-geographic trait (default) or perform the "fast" stochastic mapping to compute the expected number of events on each branch}

\item{markovjumps_total}{Whether to compute the number of events among all discrete states (default true) or not}

\item{markovjumps_pairwise}{Whether to compute the number of events between each pair of discrete states (default true) or not}

\item{format}{One of "HTML", "Latex", or "Markdown".}

\item{render_citation}{Whether the reference list will be rendered at the end of the methods document (default true; false when just showing the description in the HTML panel).}
}
\value{
a paragraph (in the selected format) describing the geographic history inference in the study.
}
\description{
Generate the description (in the selected format) of the geographic history inference in the study.
}
\examples{
tex_summarystats("Stochastic mapping (complete history, simulation-based)", TRUE, TRUE)
}
\keyword{internal}
