% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_helpr.R
\name{xml_clockratemodel}
\alias{xml_clockratemodel}
\title{Generate the average dispersal rate model block in a `BEAST` XML file.}
\usage{
xml_clockratemodel(
  discrete_trait_name,
  ctmc = F,
  clockrate_mean = 1,
  clockrate_mean_stochastic = T,
  clockrate_proposal_weight = 1,
  clockratemean_proposal_weight = 1,
  clockrate_mean_gammashaperate = 0.5
)
}
\arguments{
\item{discrete_trait_name}{name of the column containing tip states}

\item{ctmc}{Whether to specify the CTMC-rate reference prior (default prior recommended by BEAUti) on the average dispersal rate}

\item{clockrate_mean}{Mean of the Exponential prior on the average dispersal rate (relevant when an Exponential prior is used)}

\item{clockrate_mean_stochastic}{Whether the mean of the Exponential prior should be treated as a random variable with a hyperprior (default true)}

\item{clockrate_proposal_weight}{Weight of the proposal on the average dispersal rate}

\item{clockratemean_proposal_weight}{Weight of the proposal on the mean of the Exponential prior on the average dispersal rate}

\item{clockrate_mean_gammashaperate}{Shape and rate parameter (assumed to be identical) of the Gamma hyperprior of the Hierarchical Exponential prior.}
}
\value{
XML code specifying the average dispersal rate model
}
\description{
Generate the average dispersal rate model block in a `BEAST` XML file.
}
