% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_helpr.R
\name{xml_discretetraitdata}
\alias{xml_discretetraitdata}
\title{Add together two numbers}
\usage{
xml_discretetraitdata(
  states_dat,
  taxon_name,
  discrete_trait_name,
  date_name = NULL,
  lheat = 1,
  under_prior = F
)
}
\arguments{
\item{states_dat}{a data-frame object containing at least two columns: one for the name of each tip (column name specified by \code{taxon_name}), 
and the other for the trait state of each tip (column name specified by \code{discrete_trait_name})}

\item{taxon_name}{name of the column containing tip names}

\item{discrete_trait_name}{name of the column containing tip states}
}
\value{
xml code
}
\description{
Add together two numbers
}
