% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_helpr.R
\name{xml_discretetraitmodel}
\alias{xml_discretetraitmodel}
\title{Generate the discrete-geography model block in a BEAST XML file.}
\usage{
xml_discretetraitmodel(
  states_dat,
  discrete_trait_name,
  symmetry = T,
  bssvs = T,
  delta_prior = c("Poisson", "Uniform", "Beta-Binomial"),
  poisson_default = T,
  poisson_mean = 0,
  alpha_beta = 0,
  beta_beta = 0,
  rates_proposal_weight = 0,
  indicators_proposal_weight = 0,
  rootfreq_proposal_weight = 0,
  indicatorprob_proposal_weight = 0
)
}
\arguments{
\item{states_dat}{a data-frame object containing at least two columns: one for the name of each tip (column name specified by \code{taxon_name}), 
and the other for the trait state of each tip (column name specified by \code{discrete_trait_name})}

\item{discrete_trait_name}{name of the column containing tip states}

\item{symmetry}{Whether the specified geographic model is symmetric (true) or asymmetric (false)}

\item{bssvs}{Whether Bayesian Stochastic Search Variable Selection (BSSVS) is used (default true) in the discrete-geographic model.}

\item{delta_prior}{Which of the three prior options to put on \eqn{\Delta}, including:
\itemize{
  \item "Poisson": a(n offset) Poisson distribution (the default option);
  \item "Uniform": a uniform distribution between zero and the maximum \eqn{\Delta} (when all the dispersal routes exist), and;
  \item "Beta-Binomial": a Beta-Binomial distribution.
}}

\item{poisson_default}{Whether to specify the default Poisson prior in BEAUti.}

\item{poisson_mean}{The rate parameter (\eqn{\lambda}) of the Poisson distribution.}

\item{alpha_beta}{Parameter \eqn{\alpha} of the Beta-Binomial distribution.}

\item{beta_beta}{Parameter \eqn{\beta} of the Beta-Binomial distribution.}

\item{rates_proposal_weight}{Weight of the proposals on the pairwise rates of dispersal.}

\item{indicators_proposal_weight}{Weight of the proposals on the indicator variables.}

\item{rootfreq_proposal_weight}{Weight of the proposal on the root-frequency vector.}

\item{indicatorprob_proposal_weight}{Weight of the proposal on the parameter that specifies the probability of each indicator being non-zero (only relevant when \code{delta_prior} is "Beta-Binomial").}
}
\value{
XML code specifying the discrete-geography model
}
\description{
Generate the discrete-geography model block in a BEAST XML file.
}
