% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_helpr.R
\name{xml_monitors}
\alias{xml_monitors}
\title{Generate the monitor block in a `BEAST` XML file.}
\usage{
xml_monitors(
  discrete_trait_name,
  currentTree_output,
  file_name,
  bssvs = T,
  complete_history = T,
  lheat = 1,
  ctmc = T,
  clockrate_mean_stochastic = T,
  symmetry = T,
  delta_prior = c("Poisson", "Uniform", "Beta-Binomial"),
  under_prior = F,
  mcmc_samplingfreq = 1,
  ml_numstones = 0,
  ml_chainlengthperstone = 0,
  ml_samplingfreq = 0,
  ml_alphaofbeta = 0.3
)
}
\arguments{
\item{discrete_trait_name}{name of the column containing tip states}

\item{currentTree_output}{tree indices output chunk in the XML (as one of the returned value of \code{xml_treemodel})}

\item{bssvs}{Whether Bayesian Stochastic Search Variable Selection (BSSVS) is used (default true) in the discrete-geographic model.}

\item{complete_history}{Whether to perform stochastic mapping to simulate full histories of the discrete-geographic trait (default) or perform the "fast" stochastic mapping to compute the expected number of events on each branch}

\item{lheat}{Number of data clones}

\item{ctmc}{Whether to specify the CTMC-rate reference prior (default prior recommended by BEAUti) on the average dispersal rate}

\item{clockrate_mean_stochastic}{Whether the mean of the Exponential prior should be treated as a random variable with a hyperprior (default true)}

\item{symmetry}{Whether the specified geographic model is symmetric (true) or asymmetric (false)}

\item{delta_prior}{Which of the three prior options to put on \eqn{\Delta}, including:
\itemize{
  \item "Poisson": a(n offset) Poisson distribution (the default option);
  \item "Uniform": a uniform distribution between zero and the maximum \eqn{\Delta} (when all the dispersal routes exist), and;
  \item "Beta-Binomial": a Beta-Binomial distribution.
}}

\item{under_prior}{Whether to infer the joint prior distribution (default false) or the joint posterior distribution}

\item{mcmc_samplingfreq}{Every number of MCMC generations to log to the output}

\item{ml_numstones}{Number of powers ("stones") to use in the power-posterior analysis.}

\item{ml_chainlengthperstone}{Number of generations to run per stone.}

\item{ml_samplingfreq}{Every number of generations to log to the output.}

\item{ml_alphaofbeta}{Alpha of the Beta distribution that decides where to put the stones.}

\item{filename}{Name of the output file (without file extension)}
}
\value{
XML code specifying the monitor block
}
\description{
Generate the monitor block in a `BEAST` XML file.
}
