% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_helpr.R
\name{xml_phyloctmc}
\alias{xml_phyloctmc}
\title{Generate the treeLikelihood block in a `BEAST` XML file.}
\usage{
xml_phyloctmc(
  states_dat,
  discrete_trait_name,
  rootfreq_model,
  lheat = 1,
  symmetry = T,
  complete_history = T,
  do_totalcount = T,
  do_pairwisecount = T
)
}
\arguments{
\item{states_dat}{a data-frame object containing at least two columns: one for the name of each tip (column name specified by \code{taxon_name}), 
and the other for the trait state of each tip (column name specified by \code{discrete_trait_name})}

\item{discrete_trait_name}{name of the column containing tip states}

\item{rootfreq_model}{XML code specifying the root-frequency model (as one of the returned value of \code{xml_discretetraitmodel})}

\item{lheat}{Number of data clones}

\item{symmetry}{Whether the specified geographic model is symmetric (true) or asymmetric (false)}

\item{complete_history}{Whether to perform stochastic mapping to simulate full histories of the discrete-geographic trait (default) or perform the "fast" stochastic mapping to compute the expected number of events on each branch}

\item{do_totalcount}{Whether to compute the number of events among all discrete states (default true) or not.}

\item{do_pairwisecount}{Whether to compute the number of events between each pair of discrete states (default true) or not.}
}
\value{
XML code specifying the treeLikelihood block
}
\description{
Generate the treeLikelihood block in a `BEAST` XML file.
}
