% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_helpr.R
\name{xml_treemodel}
\alias{xml_treemodel}
\title{Generate the tree-model block in a `BEAST` XML file.}
\usage{
xml_treemodel(
  tree = NULL,
  treefile_name = NULL,
  tree_proposal_weight = 0,
  empiricaltree_mh = F
)
}
\arguments{
\item{tree}{A `phylo` object that contains a single tree}

\item{treefile_name}{Path of the `.trees` file that contains the distribution of trees}

\item{tree_proposal_weight}{Weight of the empirical tree operator (only becomes relevant when \code{treefile_name} is provided)}

\item{empiricaltree_mh}{Whether treating the provided distribution of trees as part of the Markov chain (default true) or averaging over the distribution uniformly.}
}
\value{
XML code specifying the tree model
}
\description{
Generate the tree-model block in a `BEAST` XML file.
}
