% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_helpr.R
\name{tex_data}
\alias{tex_data}
\title{Generate the description (in the selected format) of the data used in the study.}
\usage{
tex_data(taxa_num = 1, states_num = 1, format = c("HTML", "Latex", "Markdown"))
}
\arguments{
\item{taxa_num}{Number of taxa in the study data.}

\item{states_num}{Number of discrete states of the study trait.}

\item{format}{One of "HTML", "Latex", or "Markdown".}
}
\value{
a paragraph (in the selected format) describing the data used in the study
}
\description{
Generate the description (in the selected format) of the data used in the study.
}
\examples{
tex_data(1000, 10, "HTML")
}
\keyword{internal}
